core: move main.go to root; simplify make build/install
