+++ GNUmakefile
@@ -1,9 +1,7 @@
 .DEFAULT_GOAL := chainlink
 
-GOPATH ?= $(HOME)/go
 COMMIT_SHA ?= $(shell git rev-parse HEAD)
 VERSION = $(shell cat VERSION)
-GOBIN ?= $(GOPATH)/bin
 GO_LDFLAGS := $(shell tools/bin/ldflags)
 GOFLAGS = -ldflags "$(GO_LDFLAGS)"
 
@@ -37,25 +35,17 @@ gomodtidy: ## Run go mod tidy on all modules.
 	cd ./integration-tests && go mod tidy
 
 .PHONY: install-chainlink
-install-chainlink: chainlink ## Install the chainlink binary.
-	mkdir -p $(GOBIN)
-	rm -f $(GOBIN)/chainlink
-	cp $< $(GOBIN)/chainlink
+install-chainlink: operator-ui ## Install the chainlink binary.
+	go install $(GOFLAGS) .
 
 chainlink: operator-ui ## Build the chainlink binary.
-	go build $(GOFLAGS) -o $@ ./core/
+	go build $(GOFLAGS) .
 
 .PHONY: docker ## Build the chainlink docker image
 docker:
 	docker buildx build \
 	--build-arg COMMIT_SHA=$(COMMIT_SHA) \
 	-f core/chainlink.Dockerfile .
-
-.PHONY: chainlink-build
-chainlink-build: operator-ui ## Build & install the chainlink binary.
-	go build $(GOFLAGS) -o chainlink ./core/
-	rm -f $(GOBIN)/chainlink
-	cp chainlink $(GOBIN)/chainlink
 
 .PHONY: operator-ui
 operator-ui: ## Fetch the frontend
@@ -91,17 +81,17 @@ generate: abigen codecgen mockery ## Execute all go:generate commands.
 
 .PHONY: testdb
 testdb: ## Prepares the test database.
-	go run ./core/main.go local db preparetest
+	go run . local db preparetest
 
 .PHONY: testdb
 testdb-user-only: ## Prepares the test database with user only.
-	go run ./core/main.go local db preparetest --user-only
+	go run . local db preparetest --user-only
 
 # Format for CI
 .PHONY: presubmit
 presubmit: ## Format go files and imports.
-	goimports -w ./core
-	gofmt -w ./core
+	goimports -w .
+	gofmt -w .
 	go mod tidy
 
 .PHONY: mockery
@@ -128,7 +118,7 @@ test_need_operator_assets: ## Add blank file in web assets if operator ui has no
 
 .PHONY: config-docs
 config-docs: ## Generate core node configuration documentation
-	go run ./core/config/v2/docs/cmd/generate/main.go -o ./docs/
+	go run ./core/config/v2/docs/cmd/generate -o ./docs/
 
 .PHONY: golangci-lint
 golangci-lint: ## Run golangci-lint for all issues.
