+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/smartcontractkit/chainlink/v2/core"
+)
+
+func main() {
+	core.Main()
+}
